refactor(ingest): remove deprecated methods and warn on deprecated import (#2797)
