+++ metadata-ingestion/src/datahub/ingestion/source/ldap.py
@@ -8,7 +8,7 @@ from ldap.controls import SimplePagedResultsControl
 from datahub.configuration.common import ConfigModel, ConfigurationError
 from datahub.ingestion.api.common import PipelineContext
 from datahub.ingestion.api.source import Source, SourceReport
-from datahub.ingestion.source.metadata_common import MetadataWorkUnit
+from datahub.ingestion.api.workunit import MetadataWorkUnit
 from datahub.metadata.com.linkedin.pegasus2avro.mxe import MetadataChangeEvent
 from datahub.metadata.schema_classes import (
     CorpGroupInfoClass,
