+++ metadata-ingestion/src/datahub/integrations/airflow/__init__.py
@@ -1,3 +1,5 @@
+import warnings
+
 try:
     from datahub_provider.lineage.datahub import (
         DatahubLineageBackend as DatahubAirflowLineageBackend,
@@ -5,3 +7,8 @@ try:
 except ModuleNotFoundError:
     # Compat for older versions of Airflow.
     pass
+
+warnings.warn(
+    "importing from datahub.integrations.airflow.* is deprecated; "
+    "use datahub_provider.{hooks,operators,lineage}.datahub instead"
+)
