+++ metadata-ingestion/src/datahub/integrations/airflow/hooks.py
@@ -1,5 +1,12 @@
+import warnings
+
 from datahub_provider.hooks.datahub import (  # noqa: F401
     DatahubGenericHook,
     DatahubKafkaHook,
     DatahubRestHook,
+)
+
+warnings.warn(
+    "importing from datahub.integrations.airflow.* is deprecated; "
+    "use datahub_provider.{hooks,operators,lineage}.datahub instead"
 )
